refactor: restructured package
moved code from internal to public api package (except testutils.go)
