+++ api/rest/client_test.go
