+++ api/rest/rate.go
