+++ api/rest/reader.go
